+++ semgrep/semgrep/autofix.py
@@ -1,4 +1,3 @@
-import logging
 import re
 from pathlib import Path
 from typing import Dict
@@ -9,8 +8,9 @@ from typing import Tuple
 from semgrep.error import SemgrepError
 from semgrep.rule import Rule
 from semgrep.rule_match import RuleMatch
+from semgrep.verbose_logging import getLogger
 
-logger = logging.getLogger(__name__)
+logger = getLogger(__name__)
 
 SPLIT_CHAR = "\n"